fixed the app window, window.ship, simplified cookies and sessions, but still have audio connection issues in rooms. merging this and will focus on the last issue in rooms tomorrow

+++ app/src/os/services/ship/spaces/spaces.service.ts
@@ -80,11 +80,12 @@ export class SpacesService extends AbstractService<SpacesUpdateType> {
       switch (spacesType) {
         case 'initial':
           // TODO this DROP is here until we get the agent refactor with lastTimestamp scries
-          this.shipDB?.exec(`
-            DELETE FROM spaces_stalls;
-            DELETE FROM spaces_members;
-            DELETE FROM spaces;
-          `);
+          if (this.shipDB?.open) {
+            this.shipDB?.exec(`
+              DELETE FROM spaces_members;
+              DELETE FROM spaces;
+            `);
+          }
           this.spacesDB?.insertAll(data['initial'].spaces);
           this.spacesDB?.setCurrent(data['initial'].current.path);
           this.membersDB?.insertAll(data['initial'].membership);
